+++ changelog.md
@@ -1,6 +1,5 @@
 
 ## TODO
-	chop Op.9 No.1: no layoutMusic & midiMusic
 	track-measure lilypond section editor
 		duration checker
 	MIDI post processing
@@ -21,6 +20,8 @@
 
 
 *	2020.9.21
+
+	+	lilyTerms.ts: ContextedMusic.measureLayout added.
 
 	+	lilyInterpreter.ts: enhanced TrackContext.execute for ContextedMusic.
 	+	lilyInterpreter.ts: enhanced LilyInterpreter.execute for Variable.